pyproject.toml
[project]
name = "message-board-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "flask>=3.1.3",
    "flask-bcrypt>=1.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.2.2",
]
